separate dogecoin function
